backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
email-validator>=2.0.0
numpy>=1.24.0
